+++ internal/go_repository_tools_srcs.bzl
@@ -46,6 +46,7 @@ GO_REPOSITORY_TOOLS_SRCS = [
 	Label("//label:label.go"),
 	Label("//language:BUILD.bazel"),
 	Label("//language/go:BUILD.bazel"),
+	Label("//language/go:build_tags.go"),
 	Label("//language/go:config.go"),
 	Label("//language/go:constants.go"),
 	Label("//language/go:dep.go"),
@@ -62,6 +63,7 @@ GO_REPOSITORY_TOOLS_SRCS = [
 	Label("//language/go:package.go"),
 	Label("//language/go:resolve.go"),
 	Label("//language/go:std_package_list.go"),
+	Label("//language/go:stdlib_links.go"),
 	Label("//language/go:update.go"),
 	Label("//language/go:utils.go"),
 	Label("//language/go:work.go"),
