+++ language/go/testdata/platforms/new_tag_a.go
@@ -1,0 +1,3 @@
+//go:build amd64
+
+package platforms
